+++ test/component_test.dart
@@ -16,6 +16,19 @@ void main() {
     await widgetTester.pumpAndSettle();
     expect(acc.value, 2);
   });
+
+  testWidgets('should not rebuild on setState', (widgetTester) async {
+    final acc = Accumulator();
+    final widget = TestComponent(acc: acc);
+    await widgetTester.pumpWidget(widget);
+    await widgetTester.pumpAndSettle();
+    expect(find.byType(TestComponent), findsOneWidget);
+    final actionsState =
+        widgetTester.state(find.byType(TestComponent)) as _TestComponentState;
+    final registeredActions = actionsState.actions;
+    expect(registeredActions.containsKey(Test1Intent), isTrue);
+    expect(registeredActions.containsKey(Test2Intent), isFalse);
+  });
 }
 
 class Test1Intent extends Intent {}
@@ -34,36 +47,50 @@ class TestComponent extends StatefulWidget {
 }
 
 class _TestComponentState extends Component<TestComponent> {
-  late final test1Action = ComponentAction(_onTest1);
+  late final ComponentAction test1Action;
+  late final test2Action = ComponentAction(_onTest2);
 
   Accumulator get acc => widget.acc;
 
   @override
   void initState() {
     super.initState();
-    test1Action.attachTo(this);
+    test1Action = _onTest1.toAction().attachTo(this);
   }
 
   void _onTest1(Test1Intent intent, [BuildContext? context]) {
     test1Action.toggleEnabled();
   }
 
+  void _onTest2(Test2Intent intent, [BuildContext? context]) {
+    test2Action.toggleEnabled();
+  }
+
+  @override
+  void dispose() {
+    test1Action.detachFrom(this);
+    super.dispose();
+  }
+
   @override
   Widget build(BuildContext context) {
-    return component(
+    return withActions(
+      key: const ValueKey('actions'),
       child: MaterialApp(
         home: Scaffold(
           body: Container(
             child: Center(
               child: Column(
                 children: [
-                  Builder(builder: (context) {
-                    acc.increment(1);
-                    return ElevatedButton(
-                      onPressed: context.handler(Test1Intent()),
-                      child: Text('press'),
-                    );
-                  })
+                  Builder(
+                    builder: (context) {
+                      acc.increment(1);
+                      return ElevatedButton(
+                        onPressed: context.handler(Test1Intent()),
+                        child: Text('press'),
+                      );
+                    },
+                  )
                 ],
               ),
             ),